relocated files in IAK main folder
